backend/pyproject.toml
[project]
name = "guided-ai-coding-backend"
version = "0.1.0"
description = "Backend for AI Software Advisor"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
]
